Create 'utilities' package and add 'DateFormatter' class for date formatting accessible across activities. TT-131
